hello-world-adk/pyproject.toml
[project]
name = "hello-world-adk"
version = "0.1.0"
description = "A specialized agent that send greetings"
readme = "README.md"
requires-python = "~=3.13"
dependencies = [
    "a2a-sdk>=0.2.5",
    "google-adk>=1.2.1",
    "python-dotenv",
    "uvicorn",
    "commons",
] 

[tool.uv.workspace]
members = ["../commons"]

[tool.uv.sources]
commons = { workspace = true }